test(tls): extract per-test isolation into test-utils + #[msb_test] macro
the inline common/mod.rs helper doesn't scale as integration coverage
grows, so move the init-isolated-home logic into a dedicated workspace
crate (test-utils) and expose a proc-macro attribute (test-macros, re-
exported as test_utils::msb_test) that wraps #[tokio::test] + #[ignore] and
injects the setup call at the top of each test body.

also gate the home override on MSB_TEST_ISOLATE_HOME. when unset, the
helper is a no-op and tests run against the real ~/.microsandbox (fine for
local smoke tests serialised via -j 1). ci sets the env var so nextest can
run tests in parallel with full per-test isolation.
