+++ Cargo.toml
@@ -10,6 +10,8 @@ members = [
     "crates/migration",
     "crates/protocol",
     "crates/runtime",
+    "crates/test-macros",
+    "crates/test-utils",
     "crates/utils",
     "examples/rust/fs-read-stream",
     "examples/rust/metrics-stream",
