+++ crates/test-macros/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "test-macros"
+description = "Proc-macro for the #[msb_test] attribute. Re-exported by test-utils — depend on that instead."
+version.workspace = true
+authors.workspace = true
+repository.workspace = true
+license.workspace = true
+edition.workspace = true
+publish = false
+
+[lib]
+path = "src/lib.rs"
+proc-macro = true
+
+[dependencies]
+proc-macro2 = "1"
+quote = "1"
+syn = { version = "2", features = ["full"] }
